+++ lib/rubocop/cop/rspec_cops.rb
@@ -2,6 +2,7 @@
 
 require_relative 'rspec/capybara/current_path_expectation'
 require_relative 'rspec/capybara/feature_methods'
+require_relative 'rspec/capybara/match_style'
 require_relative 'rspec/capybara/negation_matcher'
 require_relative 'rspec/capybara/specific_actions'
 require_relative 'rspec/capybara/specific_finders'